@@ -17,8 +17,8 @@
         }
     ],
     "require": {
-        "php": "~7.2 || ^8.0",
-        "apichef/laravel-request-query-helper": "^0.5.0"
+        "php": "~7.4 || ^8.0",
+        "apichef/laravel-request-query-helper": "^1.0"
     },
     "require-dev": {
         "laravel/legacy-factories": "^1.1",
